Move hangman.setup() to package init for reusability
